feat(marketing): ship celestial landing and strip liquid glass
Replace the design-explorer landing with sectioned After Hours composition, simplify pricing shell, and remove liquid-glass marketing/nav surfaces so the celestial direction can stand alone.
