+++ src/app/(marketing)/_shared/marketing-cta.ts
@@ -1,7 +1,6 @@
 /**
  * Shared CTA styling for marketing hero + final sections.
  * Pill shape + display font (Sora via `--font-serif` → `--font-family-display`).
- * Solid/outline pills only — no liquid-glass wrapper (clips dirty on rounded-full).
  */
 export const marketingPrimaryCtaClassName =
   'group h-auto rounded-full bg-primary px-8 py-4 font-serif font-medium text-primary-foreground shadow-lg shadow-primary/25 transition-[box-shadow,transform,background-color] motion-reduce:transition-none hover:bg-primary/90 hover:shadow-xl hover:shadow-primary/30';
