--- src/app/(marketing)/_shared/marketing-glass-surface.ts
@@ -1,12 +1,0 @@
-/**
- * Shared marketing glass surface classes for cards and mock panels.
- * After Hours: parchment/plum panels — no Progress Jam white frosted washes.
- */
-export const marketingGlassCardSurface =
-  'border-panel-border/50 bg-card/50 shadow-lg backdrop-blur-xl dark:border-panel-border/60 dark:bg-panel/50';
-
-export const marketingGlassCardPopularSurface =
-  'border-primary/40 bg-card/60 shadow-lg ring-1 shadow-primary/25 ring-primary/50 backdrop-blur-xl dark:border-primary/30 dark:bg-primary/5 dark:shadow-primary/20 dark:ring-primary/70';
-
-export const marketingMockPanelSurface =
-  'overflow-hidden rounded-4xl border border-panel-border/60 bg-card/90 shadow-lg dark:border-panel-border/50';
