+++ src/app/(marketing)/landing/components/landing.module.css
@@ -1,15 +1,20 @@
-/* After Hours landing — celestial atlas motion. Scoped here, not in globals. */
+/* After Hours landing motion. */
 
-/* Scroll reveal (driven by Reveal.tsx) */
+/* Scroll reveal (driven by RevealAnimation.tsx) */
 .reveal {
-  opacity: 0;
-  transform: translateY(1.5rem);
+  opacity: 1;
+  transform: none;
   transition:
     opacity 0.7s cubic-bezier(0.22, 1, 0.36, 1),
     transform 0.7s cubic-bezier(0.22, 1, 0.36, 1);
 }
 
-.revealVisible {
+.reveal[data-reveal-state='observing'] {
+  opacity: 0;
+  transform: translateY(1.5rem);
+}
+
+.reveal[data-reveal-state='visible'] {
   opacity: 1;
   transform: none;
 }
@@ -26,52 +31,6 @@
   }
   50% {
     transform: translateY(-0.5rem);
-  }
-}
-
-/* Constellation route: draws in once its Reveal wrapper becomes visible */
-.routePath {
-  stroke-dasharray: 1;
-  stroke-dashoffset: 1;
-}
-
-.revealVisible .routePath {
-  animation: drawRoute 1.8s cubic-bezier(0.4, 0, 0.2, 1) 0.2s forwards;
-}
-
-@keyframes drawRoute {
-  to {
-    stroke-dashoffset: 0;
-  }
-}
-
-/* Star nodes pop after the route draws under them */
-.routeNode {
-  opacity: 0;
-  transform-origin: center;
-  transform: scale(0);
-}
-
-.revealVisible .routeNode {
-  animation: popNode 0.5s cubic-bezier(0.34, 1.56, 0.64, 1) forwards;
-}
-
-.revealVisible .routeNodeDelay1 {
-  animation-delay: 0.5s;
-}
-
-.revealVisible .routeNodeDelay2 {
-  animation-delay: 1.1s;
-}
-
-.revealVisible .routeNodeDelay3 {
-  animation-delay: 1.7s;
-}
-
-@keyframes popNode {
-  to {
-    opacity: 1;
-    transform: scale(1);
   }
 }
 
@@ -101,17 +60,6 @@
 
   .drift,
   .pulse {
-    animation: none;
-  }
-
-  .routePath {
-    stroke-dashoffset: 0;
-    animation: none;
-  }
-
-  .routeNode {
-    opacity: 1;
-    transform: scale(1);
     animation: none;
   }
 }
