+++ src/app/(marketing)/pricing/components/Pricing.module.css
@@ -1,6 +1,4 @@
-/* After Hours pricing page shell — celestial backdrop + hero entrance.
-   Card / period-tab chrome lives in
-   `src/app/(marketing)/_shared/after-hours-pricing-cards.module.css`. */
+/* After Hours pricing page shell — celestial backdrop + hero entrance. */
 
 .shell {
   position: relative;
@@ -11,8 +9,7 @@
   background: var(--background);
 }
 
-/* Hero entrance — overline rises first, then the headline word by word,
-   then the subline settles in. */
+/* Hero entrance — overline rises first, then the headline and subline. */
 
 .heroOverline {
   font-family: var(--font-family-display);
@@ -25,13 +22,11 @@
   animation: heroRise 0.6s cubic-bezier(0.22, 1, 0.36, 1) forwards;
 }
 
-.heroWord {
-  display: inline-block;
+.heroHeading {
   opacity: 0;
   transform: translateY(0.5em);
   filter: blur(4px);
-  animation: wordRise 0.7s cubic-bezier(0.22, 1, 0.36, 1)
-    calc(0.2s + var(--word-index, 0) * 90ms) forwards;
+  animation: heroRise 0.7s cubic-bezier(0.22, 1, 0.36, 1) 0.2s forwards;
 }
 
 .heroSubline {
@@ -66,14 +61,6 @@
   }
 }
 
-@keyframes wordRise {
-  to {
-    opacity: 1;
-    transform: none;
-    filter: none;
-  }
-}
-
 @media (max-width: 900px) {
   .shell {
     padding-top: 0;
@@ -87,7 +74,7 @@
     animation: none;
   }
 
-  .heroWord {
+  .heroHeading {
     opacity: 1;
     transform: none;
     filter: none;
