--- src/app/(marketing)/pricing/components/PricingFinalCta.tsx
@@ -1,17 +1,0 @@
-'use client';
-
-import { MarketingCta } from '@/app/(marketing)/_shared/MarketingCta';
-import { useId } from 'react';
-
-export function PricingFinalCta() {
-  const headingId = useId();
-
-  return (
-    <MarketingCta
-      headingId={headingId}
-      title='Start learning for free'
-      description='Create your first plan at no cost. Upgrade when you need more capacity.'
-      buttonLabel='Get started free'
-    />
-  );
-}
