+++ src/app/globals.css
@@ -164,7 +164,6 @@
     ),
     linear-gradient(140deg, #180d18 0%, #2b1728 52%, #351b30 100%);
   --app-background-size: auto, auto, auto;
-
 }
 
 @theme inline {
@@ -214,8 +213,6 @@
   --font-sans: var(--font-sans);
   --font-mono: var(--font-mono);
   --font-serif: var(--font-serif);
-  /* Alias for Sora display stack; `font-serif` remains the legacy utility name. */
-  --font-display: var(--font-family-display);
 
   --radius-sm: calc(var(--radius) * 0.75);
   --radius-md: calc(var(--radius) * 0.875);
@@ -302,88 +299,6 @@
     line-height: 1.5;
     letter-spacing: 0;
     color: var(--muted-foreground);
-  }
-
-  /*
-   * Marketing Typography - Large scale for landing/marketing pages
-   * Major Third (1.25x) scale with tight letter-spacing
-   */
-  .marketing-h1 {
-    font-family: var(--font-family-display);
-    font-weight: 600;
-    font-size: 3.0625rem; /* 49px */
-    line-height: 1.1;
-    letter-spacing: -0.03em;
-    text-wrap: balance;
-  }
-
-  .marketing-h2 {
-    font-family: var(--font-family-display);
-    font-weight: 600;
-    font-size: 2.4375rem; /* 39px */
-    line-height: 1.15;
-    letter-spacing: -0.025em;
-    text-wrap: balance;
-  }
-
-  .marketing-h3 {
-    font-family: var(--font-family-display);
-    font-weight: 600;
-    font-size: 1.9375rem; /* 31px */
-    line-height: 1.2;
-    letter-spacing: -0.02em;
-    text-wrap: balance;
-  }
-
-  .marketing-h4 {
-    font-family: var(--font-family-display);
-    font-weight: 600;
-    font-size: 1.5625rem; /* 25px */
-    line-height: 1.25;
-    letter-spacing: -0.015em;
-    text-wrap: balance;
-  }
-
-  /* Card titles inside marketing sections (features, steps, problem/solution) */
-  .marketing-card-title {
-    font-family: var(--font-family-heading);
-    font-weight: var(--font-weight-heading);
-    font-size: 1.25rem; /* 20px — aligns with text-xl */
-    line-height: 1.3;
-    letter-spacing: -0.015em;
-    text-wrap: balance;
-    color: var(--foreground);
-  }
-
-  .marketing-subtitle {
-    font-family: var(--font-family-base);
-    font-weight: var(--font-weight-base);
-    font-size: 1.25rem; /* 20px */
-    line-height: 1.5;
-    letter-spacing: 0;
-    color: var(--muted-foreground);
-  }
-
-  /* Responsive scaling for marketing typography */
-  @media (max-width: 640px) {
-    .marketing-h1 {
-      font-size: 2.4375rem; /* 39px on mobile */
-    }
-    .marketing-h2 {
-      font-size: 1.9375rem; /* 31px on mobile */
-    }
-    .marketing-h3 {
-      font-size: 1.5625rem; /* 25px on mobile */
-    }
-    .marketing-h4 {
-      font-size: 1.25rem; /* 20px on mobile */
-    }
-    .marketing-card-title {
-      font-size: 1.125rem; /* 18px on mobile */
-    }
-    .marketing-subtitle {
-      font-size: 1rem; /* 16px on mobile */
-    }
   }
 }
