--- src/components/shared/liquid-glass/index.ts
@@ -1,10 +1,0 @@
-export { LiquidGlass } from './LiquidGlass';
-export { LiquidGlassLayer } from './LiquidGlassLayer';
-export {
-  DEFAULT_LIQUID_GLASS_PHYSICS,
-  MARKETING_CTA_PHYSICS,
-  MARKETING_HEADER_PHYSICS,
-  PRICING_HEADER_PHYSICS,
-  resolveLiquidGlassPhysics,
-  type LiquidGlassPhysics,
-} from './types';
