+++ tests/unit/app/pricing/page.spec.tsx
@@ -3,34 +3,21 @@ import { render, screen } from '@testing-library/react';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 
 const mocks = vi.hoisted(() => ({
-  clerkPlansMock: vi.fn(),
   getOptionalCheckoutBillingSignatureMock: vi.fn(),
   pricingTableMock: vi.fn(),
   shouldUseClerkUiMock: vi.fn(() => true),
 }));
 
-vi.mock(
-  '@/app/(marketing)/_shared/after-hours-pricing-cards.module.css',
-  () => ({ default: { checkoutMount: 'checkoutMount' } }),
-);
-
-vi.mock(
-  '@/app/(marketing)/pricing/components/PricingAfterHours.module.css',
-  () => ({
-    default: {
-      heroOverline: 'heroOverline',
-      heroSubline: 'heroSubline',
-      heroWord: 'heroWord',
-      shell: 'shell',
-    },
-  }),
-);
-
 vi.mock('@/app/(marketing)/_shared/star-field.module.css', () => ({
+  default: { star: 'star' },
+}));
+
+vi.mock('@/app/(marketing)/pricing/components/Pricing.module.css', () => ({
   default: {
-    stars: 'stars',
-    starsFar: 'starsFar',
-    starsNear: 'starsNear',
+    heroHeading: 'heroHeading',
+    heroOverline: 'heroOverline',
+    heroSubline: 'heroSubline',
+    shell: 'shell',
   },
 }));
 
@@ -48,8 +35,6 @@ vi.mock('@clerk/nextjs', () => ({
     mocks.pricingTableMock(props);
     return <div data-testid='clerk-pricing-table' />;
   },
-  useAuth: () => ({ isLoaded: true, userId: null }),
-  useClerk: () => ({ billing: { getPlans: mocks.clerkPlansMock } }),
 }));
 
 async function renderPricingPage(): Promise<void> {
@@ -62,7 +47,6 @@ async function renderPricingPage(): Promise<void> {
 describe('PricingPage', () => {
   beforeEach(() => {
     vi.resetAllMocks();
-    mocks.clerkPlansMock.mockResolvedValue({ data: [] });
     mocks.getOptionalCheckoutBillingSignatureMock.mockResolvedValue(
       'free|active||0',
     );
@@ -83,11 +67,6 @@ describe('PricingPage', () => {
       }),
     ).toBeVisible();
     expect(screen.getByText(/chart your course/i)).toBeVisible();
-    expect(
-      screen.getByRole('group', { name: /billing period/i }),
-    ).toBeVisible();
-    expect(screen.getByRole('button', { name: /^monthly$/i })).toBeVisible();
-    expect(screen.getByRole('button', { name: /^yearly$/i })).toBeVisible();
     expect(screen.getByTestId('clerk-pricing-table')).toBeVisible();
     expect(mocks.pricingTableMock).toHaveBeenCalledWith(
       expect.objectContaining({
@@ -102,9 +81,6 @@ describe('PricingPage', () => {
     await renderPricingPage();
 
     expect(screen.queryByTestId('clerk-pricing-table')).not.toBeInTheDocument();
-    expect(
-      screen.queryByRole('tablist', { name: /billing period/i }),
-    ).not.toBeInTheDocument();
     expect(screen.getByRole('alert')).toBeVisible();
     expect(screen.getByText(/local product testing mode/i)).toBeVisible();
     expect(screen.getByText(/billing:clerk:fixture/i)).toBeVisible();
